model_01_segmentation/requirements.txt
# Model 01 — NBA Player Segmentation
# Python 3.11+
# Install with: pip install -r requirements.txt

nba_api==1.11.4
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
umap-learn>=0.5
plotly>=5.0
matplotlib>=3.7
seaborn>=0.12
jupyter>=1.0
ipykernel>=6.0
